requirements.txt
fastapi
uvicorn
python-dotenv
serpapi
google-play-scraper
pandas
nltk
transformers
torch
scikit-learn
pydantic
